LPD-29313 Move to pages since they don't have tests
