--- modules/test/playwright/tests/portal-workflow-task-web/test.properties
@@ -1,5 +1,0 @@
-##
-## Testray
-##
-
-    testray.main.component.name=Workflow Task
